[fix] makefile app_version logic
We are fixing the logic to extract the APP_Version.

Specifically, we are checking if we are in protected branch, if yes use we use `latest` as app version.
Else we are checking if we are in a release tag, if yes we are using tag as app version
Else we use `dev-sha` as app version

Signed-off-by: Akis Maziotis <akis.maziotis@mattermost.com>