[feat] Makefile APP_VERSION assigment revisit
There maybe requirment to set APP_VERSION at CI time.
For example if tagging from protected branch to release, then this will happen:
`if [ $(PROTECTED_BRANCH) = $(CURRENT_BRANCH) ];`
setting the APP_VERSION=$(PROTECTED_BRANCH)-$(APP_COMMIT)

In that case, we want to overwrite the APP_VERSION to be equal to the tag name.

Therefore we are changing the variable assignment to "?=" to be able to cover that case